requirements.txt
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
